Put websocket support behind SPI.
